+++ interfacev2/transaction.go
@@ -41,7 +41,7 @@ type Transaction interface {
 	RecoverChainState(ChainStateOperation) error
 
 	// fee
-	FeePurity() uint64 // fee purity
+	FeePurity() uint32 // fee purity
 
 	// get something
 	GetAddress() fields.Address
